fw_meta: Split up fw_meta.rs
To facilitate platform abstraction, split fw_meta.rs into:

- kernel/src/utils/fw_meta.rs: utility data structures/functions
- kernel/src/platform/snp_fw.rs: SEV-specific platform operations

No functional change intended.
