+++ kernel/src/config.rs
@@ -12,9 +12,9 @@ use crate::acpi::tables::{load_acpi_cpu_info, ACPICPUInfo};
 use crate::address::PhysAddr;
 use crate::error::SvsmError;
 use crate::fw_cfg::FwCfg;
-use crate::fw_meta::{parse_fw_meta_data, SevFWMetaData};
 use crate::igvm_params::IgvmParams;
 use crate::mm::{PerCPUPageMappingGuard, PAGE_SIZE, SIZE_1G};
+use crate::platform::snp_fw::{parse_fw_meta_data, SevFWMetaData};
 use crate::serial::SERIAL_PORT;
 use crate::utils::MemoryRegion;
 use alloc::vec::Vec;
